requirements.txt
matplotlib
numpy
openpyxl
pandas
psycopg[binary]
python-dateutil
python-dotenv
